exercise_one/user_service/requirements.txt
Flask
requests
SQLAlchemy
flask_sqlalchemy
psycopg2-binary
pika
python-dotenv